Avoid WMF-custom "warning: Unknown link" for ES6 built-ins
Allowing these to be linked to MDN is useful, but should not be
mandatory. Also expand to list to include classes and types that
JSDuck recognised by default, as well as some more recent ones
from ES6, such as the 'symbol' type, and Promise/Map/Set objects.

Also, use linkText (which is html-safe) as contents instead of
longname. Output is the same since all known matches are in the
regexp literally, but makes it more obvious that it is safe.

Change-Id: I38d1f2f75aa14e00467e7fdd6e5d60fde9c4716f