tables: Add horizontal scroll
Add a wrapper div around params and props tables and set styles
to prevent the div from being wider than its container, and to
add horizontal scroll when its content exceeds the max-width.

To improve the UI, the outer border of the table was moved to
this div. As a result, we can't rely on `border-collapse` to
handle overlapping borders, and must instead set only the inner
border of table cells.

Bug: T359113
Change-Id: Iee253f414a51c4890950e6deb8f20a98475d432d