js: Fix redirecting for links to methods and properties
Prior to this change, old links to methods and properties redirected to a 404 page
(e.g. https://doc.wikimedia.org/mediawiki-core/master/js/#!/api/mw.Uri-method-clone; more examples
at T352320#9776297). With this change, the user is redirected correctly for both static and
instance methods and properties. Other possible postfixes starting with "-" are removed (at least
the user will be redirected to an existing page and not 404).