Merge prefixMap and prefixMapIgnore
Now that prefixMapIgnore is an object, we can just merge it
with prefixMap. prefixMap values can now be a boolean where
`false` unregisters the prefix completely, allowing shorter
sub-prefixes to be registered, and `true` registers the
prefix as a local one but skips using external links for it.

The `true` value can be used, for example, in OOUI to indicate
that 'OO.ui.' is always defined locally, and not to fall through
to the 'OO.' prefix.

Change-Id: Ib72cbfeeb64c8575d0609e1e74a74560c1f14172