js: Fix redirecting for links to methods, properties, and events
Prior to this change, old links to methods and properties redirected to a 404 page
(e.g. https://doc.wikimedia.org/mediawiki-core/master/js/#!/api/mw.Uri-method-clone; more examples
at T352320#9776297). With this change, the user is redirected correctly for both static and
instance methods and properties, as well as for events. Other possible postfixes starting with "-"
are removed (at least the user will be redirected to an existing page and not 404).

This still leaves out some cases, e.g.:
* https://doc.wikimedia.org/mediawiki-core/master/js/#!/api/mw.util-method-getUrl redirects to a
  page which duplicates
  https://doc.wikimedia.org/mediawiki-core/master/js/module-mediawiki.util.html and has no
  content.
* https://doc.wikimedia.org/oojs-ui/master/js/#!/api/OO.ui.MessageWidget-cfg-type redirects to the
  right page, but not the constructor section where the config parameters reside.

Bug: T352320
Change-Id: If189bfd612ee0b35b39944436db045edbdeb319f