Fix hidden search results list
I couldn't replicate this on localhost, but in production,
the .search_results list is covered up by the .page div.

This patch changes the .search_results z-index from auto (0)
to 1. Making this change manually in my browser fixes the
issue in production.

Couldn't replicate on localhost, but worth a try.

Bug: T364968
Change-Id: I8b379cc5b6fe3d2a8477c844b51bdb297404a66e